move DoorFlag, requiired keywords
