+++ tests/conftest.py
@@ -13,8 +13,8 @@ from cibo.models.data.npc import Npc as NpcData
 from cibo.models.data.player import Player
 from cibo.models.description import EntityDescription, RoomDescription
 from cibo.models.direction import Direction
-from cibo.models.door import Door, DoorFlag
-from cibo.models.flag import RoomFlag
+from cibo.models.door import Door
+from cibo.models.flag import DoorFlag, RoomFlag
 from cibo.models.item import Item
 from cibo.models.npc import Npc
 from cibo.models.region import Region
